+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 `hakyll-shortcut-links` uses [PVP Versioning][1].
 The changelog is available [on GitHub][2].
 
-## Unreleased
+## 0.1.0.0 — Feb 29, 2020
 
 * [#16](https://github.com/kowainik/hakyll-shortcut-links/issues/16):
   Support GHC-8.8.2. Update `hakyll` and `pandoc` bounds.